requirements.txt
alabaster==0.7.10
altgraph==0.14
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
docutils==0.13.1
idna==2.5
imagesize==0.7.1
Jinja2==2.9.6
macholib==1.8
MarkupSafe==1.0
modulegraph==0.15
py2app==0.14
Pygments==2.2.0
PyQt5==5.9
pyserial==3.4
pytz==2017.2
requests==2.18.2
sip==4.19.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
urllib3==1.22
